feat: refactoring-4 created more components
