+++ src/index.css
@@ -50,120 +50,6 @@ header nav ul li.active {
   color: #ffffff;
 }
 
-.search__mark_avatar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 24px;
-  width: 555px;
-  max-width: 555px;
-}
-
-.mark__avatar__imgs {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 24px;
-}
-
-.search {
-  display: flex;
-  justify-content: end;
-  align-items: center;
-  gap: 4px;
-  width: 445px;
-  padding: 7px 10px;
-  background-color: #161615;
-  border-radius: 4px;
-  cursor: pointer;
-}
-
-.search__input {
-  width: 100%;
-  border: none;
-  color: #ffffff;
-  background-color: transparent;
-}
-
-.search__close__btn {
-  width: 16px;
-  height: 16px;
-}
-
-.search__btn {
-  display: none;
-  justify-content: center;
-  align-items: center;
-  font-size: 14px;
-  line-height: 19.6px;
-  color: #ffffff;
-  background-color: transparent;
-  border: none;
-  cursor: pointer;
-}
-
-.show__search__btn {
-  display: flex;
-}
-
-.mark {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 24px;
-  height: 24px;
-  cursor: pointer;
-}
-
-.avatar {
-  width: 32px;
-  height: 32px;
-  border-radius: 100px;
-  cursor: pointer;
-}
-
-.dropdown {
-  display: none;
-}
-
-main {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  padding: 24px 40px 40px 40px;
-  color: #ffffff;
-  background-color: #161615;
-}
-
-main h2 {
-  font-size: 20px;
-  line-height: 28px;
-  letter-spacing: 1px;
-}
-
-.movies__list {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 20px;
-}
-
-.noresult__wrapper {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  font-size: 82px;
-  color: #ffffff;
-}
-
-.noresult__message {
-  font-size: 62px;
-  color: #737373;
-}
-
-footer {
-  display: none;
-}
-
 @media (max-width: 1055px) {
   header nav ul {
     display: none;
@@ -188,90 +74,6 @@ footer {
   .hide__logo {
     display: none;
   }
-
-  .search__mark_avatar {
-    justify-content: end;
-    width: 100%;
-    max-width: 100%;
-  }
-
-  .hide__imgs {
-    display: none;
-  }
-
-  .search {
-    display: none;
-    justify-content: end;
-    width: 100%;
-  }
-
-  .show__search {
-    display: flex;
-  }
-
-  .search__close__btn {
-    width: 16px;
-    height: 16px;
-  }
-
-  .show__search__btn {
-    display: flex;
-  }
-
-  main {
-    padding: 24px 16px;
-  }
-
-  .movies__list {
-    justify-content: space-evenly;
-  }
-
-  .noresult__wrapper {
-    font-size: 62px;
-  }
-
-  .noresult__message {
-    font-size: 42px;
-  }
-
-  footer {
-    display: block;
-  }
-
-  footer nav {
-    padding: 14px 0;
-    background-color: #080808;
-  }
-
-  footer nav ul {
-    display: flex;
-    justify-content: space-evenly;
-    align-items: center;
-  }
-
-  .footer__btn {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 2px;
-    padding: 0;
-    color: #ffffff;
-    background-color: transparent;
-    opacity: 0.5;
-    border: none;
-    cursor: pointer;
-  }
-
-  .footer__btn p {
-    font-size: 8px;
-    letter-spacing: 0.5px;
-  }
-
-  footer .active {
-    color: #e6e6e6;
-    opacity: 1;
-  }
 }
 
 @media (max-width: 500px) {
@@ -294,74 +96,5 @@ footer {
 
   header nav ul {
     display: none;
-  }
-
-  .search__mark_avatar {
-    justify-content: end;
-    width: 100%;
-  }
-
-  .mark__avatar__imgs {
-    gap: 0;
-  }
-
-  .hide__imgs {
-    display: none;
-  }
-
-  .search {
-    display: none;
-    justify-content: end;
-    width: 100%;
-  }
-
-  .show__search {
-    display: flex;
-  }
-
-  .search__close__btn {
-    width: 16px;
-    height: 16px;
-  }
-
-  .show__search__btn {
-    display: flex;
-  }
-
-  .mark {
-    display: none;
-  }
-
-  .dropdown {
-    display: block;
-  }
-
-  main {
-    padding: 24px 16px;
-  }
-
-  main h2 {
-    font-size: 16px;
-    line-height: 22.4px;
-  }
-
-  .movies__list {
-    gap: 16px;
-  }
-
-  .noresult__wrapper {
-    font-size: 62px;
-  }
-
-  .noresult__message {
-    font-size: 42px;
-  }
-
-  footer {
-    display: block;
-  }
-
-  footer nav {
-    padding: 6px 0;
   }
 }
